requirements.txt
adamp==0.3.0
spacy==2.3.2
networkx==2.4
scikit-learn==0.23.1
matplotlib==3.2.2
torchtext==0.6.0
lmdb==0.98
numpy==1.18.5
tqdm==4.46.1
requests==2.24.0
pandas==1.0.5
hydra-core==0.11.3
hydra-colorlog==0.1.4
pyarrow==0.16.0
higher==0.2.1
GPUtil==1.4.0
dgl-cu100==0.5.1
omegaconf==1.4.1
seaborn
xlrd